+++ src/lang/lang_errors.ml
@@ -51,13 +51,13 @@ let report lexbuf f =
   let print_error idx error =
     flush_all () ;
     let pos =
-      let start = lexbuf.Sedlexing_compat.lex_curr_p in
-      let buf = Sedlexing_compat.Utf8.lexeme lexbuf in
+      let start = snd (Sedlexing.lexing_positions lexbuf) in
+      let buf = Sedlexing.Utf8.lexeme lexbuf in
       Printf.sprintf "%sine %d, char %d%s"
         (if start.Lexing.pos_fname="" then "L" else
            Printf.sprintf "File %S, l" start.Lexing.pos_fname)
         start.Lexing.pos_lnum
-        (1+start.Lexing.pos_cnum-start.Lexing.pos_bol)
+        (start.Lexing.pos_cnum-start.Lexing.pos_bol)
         (if buf = "" then "" else (Printf.sprintf " before %S" buf)) ;
     in
     error_header idx pos;
@@ -148,5 +148,7 @@ let report lexbuf f =
         error_header 12 pos;
         Format.printf "Unsupported format!@ You must be missing an optional dependency.@]@.";
         raise Error
+      | Sedlexing.MalFormed ->
+        print_error 13 "Malformed file."
       | End_of_file -> raise End_of_file
       | e -> print_error (-1) "Unknown error" ; raise e